Minecraft/BGM/parser/requirements.txt
anvil-parser